+++ core/src/executive/executive_tests.rs
@@ -5,6 +5,7 @@
 use super::{executive::*, internal_contract::*, Executed, ExecutionError};
 use crate::{
     evm::{Factory, FinalizationResult, VMType},
+    executive::ExecutionOutcome,
     machine::Machine,
     parameters::staking::*,
     state::{CleanupMode, CollateralCheckResult, Substate},
@@ -465,18 +466,22 @@ fn test_not_enough_cash() {
             &spec,
             &internal_contract_map,
         );
-        let mut nonce_increased = false;
-        ex.transact(&t, &mut nonce_increased)
+        ex.transact(&t).unwrap()
     };
 
     match res {
-        Err(ExecutionError::NotEnoughCash {
-            required,
-            got,
-            actual_gas_cost,
-        }) if required == U512::from(100_018)
+        ExecutionOutcome::ExecutionErrorBumpNonce(
+            ExecutionError::NotEnoughCash {
+                required,
+                got,
+                actual_gas_cost,
+                max_storage_limit_cost,
+            },
+            _executed,
+        ) if required == U512::from(100_018)
             && got == U512::from(100_017)
-            && correct_cost == actual_gas_cost =>
+            && correct_cost == actual_gas_cost
+            && max_storage_limit_cost.is_zero() =>
         {
             ()
         }
@@ -965,7 +970,6 @@ fn test_commission_privilege() {
         )
         .unwrap();
 
-    let mut nonce_increased: bool = false;
     let tx = Transaction {
         nonce: 0.into(),
         gas_price: U256::from(1),
@@ -986,7 +990,9 @@ fn test_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(gas_used, U256::from(58_024));
@@ -1076,7 +1082,9 @@ fn test_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(gas_used, U256::from(58_024));
@@ -1115,7 +1123,9 @@ fn test_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(gas_used, U256::from(58_024));
@@ -1154,7 +1164,9 @@ fn test_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(gas_used, U256::from(58_024));
@@ -1207,7 +1219,9 @@ fn test_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(gas_used, U256::from(58_024));
@@ -1253,7 +1267,9 @@ fn test_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(gas_used, U256::from(58_024));
@@ -1315,14 +1331,13 @@ fn test_storage_commission_privilege() {
         .unwrap();
 
     // simple call to create a storage entry
-    let mut nonce_increased: bool = false;
     let tx = Transaction {
         nonce: 0.into(),
         gas_price: U256::from(1),
         gas: U256::from(100_000),
         value: *COLLATERAL_PER_STORAGE_KEY,
         action: Action::Call(address),
-        storage_limit: U256::from(64),
+        storage_limit: U256::from(BYTES_PER_STORAGE_KEY),
         epoch_height: 0,
         chain_id: 0,
         data: vec![],
@@ -1341,11 +1356,13 @@ fn test_storage_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
     assert_eq!(storage_collateralized.len(), 1);
     assert_eq!(storage_collateralized[0].address, sender.address());
-    assert_eq!(storage_collateralized[0].amount, 64);
+    assert_eq!(storage_collateralized[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(storage_released.len(), 0);
 
     state
@@ -1417,7 +1434,10 @@ fn test_storage_commission_privilege() {
     );
     state.discard_checkpoint();
     assert_eq!(substate.storage_collateralized.len(), 1);
-    assert_eq!(substate.storage_collateralized[&sender.address()], 2 * 64);
+    assert_eq!(
+        substate.storage_collateralized[&sender.address()],
+        2 * BYTES_PER_STORAGE_KEY
+    );
     assert_eq!(
         *state.total_storage_tokens(),
         *COLLATERAL_PER_STORAGE_KEY * U256::from(3)
@@ -1455,7 +1475,7 @@ fn test_storage_commission_privilege() {
         gas: U256::from(100_000),
         value: U256::from(0),
         action: Action::Call(address),
-        storage_limit: U256::from(64),
+        storage_limit: U256::from(BYTES_PER_STORAGE_KEY),
         epoch_height: 0,
         chain_id: 0,
         data: vec![],
@@ -1474,15 +1494,17 @@ fn test_storage_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(storage_collateralized.len(), 1);
     assert_eq!(storage_collateralized[0].address, caller3.address());
-    assert_eq!(storage_collateralized[0].amount, 64);
+    assert_eq!(storage_collateralized[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(storage_released.len(), 1);
     assert_eq!(storage_released[0].address, sender.address());
-    assert_eq!(storage_released[0].amount, 64);
+    assert_eq!(storage_released[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(gas_used, U256::from(26_017));
     assert_eq!(
         state.sponsor_balance_for_collateral(&address).unwrap(),
@@ -1528,7 +1550,7 @@ fn test_storage_commission_privilege() {
         gas: U256::from(100_000),
         value: U256::from(0),
         action: Action::Call(address),
-        storage_limit: U256::from(64),
+        storage_limit: U256::from(BYTES_PER_STORAGE_KEY),
         epoch_height: 0,
         chain_id: 0,
         data: vec![],
@@ -1547,15 +1569,17 @@ fn test_storage_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(storage_collateralized.len(), 1);
     assert_eq!(storage_collateralized[0].address, address);
-    assert_eq!(storage_collateralized[0].amount, 64);
+    assert_eq!(storage_collateralized[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(storage_released.len(), 1);
     assert_eq!(storage_released[0].address, caller3.address());
-    assert_eq!(storage_released[0].amount, 64);
+    assert_eq!(storage_released[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(gas_used, U256::from(26_017));
     assert_eq!(
         state.balance(&caller1.address()).unwrap(),
@@ -1619,7 +1643,7 @@ fn test_storage_commission_privilege() {
         gas: U256::from(100_000),
         value: U256::from(0),
         action: Action::Call(address),
-        storage_limit: U256::from(64),
+        storage_limit: U256::from(BYTES_PER_STORAGE_KEY),
         epoch_height: 0,
         chain_id: 0,
         data: vec![],
@@ -1638,15 +1662,17 @@ fn test_storage_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(storage_collateralized.len(), 1);
     assert_eq!(storage_collateralized[0].address, caller2.address());
-    assert_eq!(storage_collateralized[0].amount, 64);
+    assert_eq!(storage_collateralized[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(storage_released.len(), 1);
     assert_eq!(storage_released[0].address, address);
-    assert_eq!(storage_released[0].amount, 64);
+    assert_eq!(storage_released[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(gas_used, U256::from(26_017));
     assert_eq!(
         state.balance(&caller2.address()).unwrap(),
@@ -1716,7 +1742,10 @@ fn test_storage_commission_privilege() {
         *COLLATERAL_PER_STORAGE_KEY * U256::from(1),
     );
     assert_eq!(substate.storage_released.len(), 1);
-    assert_eq!(substate.storage_released[&sender.address()], 64);
+    assert_eq!(
+        substate.storage_released[&sender.address()],
+        BYTES_PER_STORAGE_KEY
+    );
     assert_eq!(
         *state.total_storage_tokens(),
         *COLLATERAL_PER_STORAGE_KEY * U256::from(2)
@@ -1740,7 +1769,7 @@ fn test_storage_commission_privilege() {
         gas: U256::from(100_000),
         value: U256::from(0),
         action: Action::Call(address),
-        storage_limit: U256::from(64),
+        storage_limit: U256::from(BYTES_PER_STORAGE_KEY),
         epoch_height: 0,
         chain_id: 0,
         data: vec![],
@@ -1759,15 +1788,17 @@ fn test_storage_commission_privilege() {
         &spec,
         &internal_contract_map,
     )
-    .transact(&tx, &mut nonce_increased)
+    .transact(&tx)
+    .unwrap()
+    .successfully_executed()
     .unwrap();
 
     assert_eq!(storage_collateralized.len(), 1);
     assert_eq!(storage_collateralized[0].address, caller1.address());
-    assert_eq!(storage_collateralized[0].amount, 64);
+    assert_eq!(storage_collateralized[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(storage_released.len(), 1);
     assert_eq!(storage_released[0].address, caller2.address());
-    assert_eq!(storage_released[0].amount, 64);
+    assert_eq!(storage_released[0].amount, BYTES_PER_STORAGE_KEY);
     assert_eq!(gas_used, U256::from(26_017));
     assert_eq!(
         state.balance(&caller1.address()).unwrap(),